Scan & Savour
This is my college PBL(Project Based Learning) project of IV Semester. In this project, I have developed a website by which anyone generate QR menu for his shop. It will help them to manage in their work and orders efficiently. This project is developed with HTML, CSS and JavaScript, Node.js, Express.js and MongoDB.